[android] remove usage of StudioFlags class in android gradle-dsl module
Android gradle-dsl module should be Android plugin agnostic, but with the latest
changes in AS Hedghehog Beta 5 the usage of StudioFlags class is introduced to hide
`Enable support for gradle.build.toml in PSD and Assistants` studio feature under
feature flag.

Instead of StudioFlags class, the feature flags was moved to Registry.

GitOrigin-RevId: c12b5fab83d690e679be0cf5a7ee83687fec5f7a