Add quickfixes to support removed error on JDK 21
Using source/target 6/7 when compiling with JDK21 results in support
removed error. AGP now generates its own error for this which
results in no usual javac lines in the output which we were parsing
previously. Actually no output is generated apart from final failure
message.

This means we need to implement issue checker to catch this case. But
we can not emit an issue from issueChecker's 'check' function - this
ends up to be a root level issue when we want it to be a 'compilation
issue' attached to javac task.
To do this we emit the issue from 'consumeBuildOutputFailureMessage'
which comes from the parsed failure message including the task name that
failed.

Additionally, we need to add notification extension that would filter
out this exception from being converted to Failure object and shown in
BOW as duplicate on root level. This is recommended for errors from
'compiler' as these should be processed by parsers.

Bug: 337007673
Test: JavaLanguageLevelDeprecationOutputParserTest
Change-Id: I136fe3744d372679c61f474ed8ad4fc4673fb35a

GitOrigin-RevId: 7f91bb3d1d66947173987005a0055d76cf17a0cd