Add file context to PromptBuilder
This CL adds a new method to the buildPrompt DSL: context

Inside a context block you can specify one or more files to use as
context. You can specify whether the files are currently open, and
whether there is a selection in each file. The files must of course
comply with aiexclude.

This also adds new flags for current file context and open files context
and removes old context flags that are no longer needed.

Bug: 331957364
Test: included
Change-Id: Ic81886975205484a7825164efed2d0a5f6fa2d3b

GitOrigin-RevId: 9697a16fff9bb52aa2a2f43beffd607e1910577f