Instatiate both test preview inspections in tests
This needs to be done explicitly since the IntelliJ test framework does
not allow the same inspection tool class to be registered twice in the
2024.1.1 platform. Without this change, only the kotlin version gets
registered in tests causing java test failures.

For a long term fix, we should convert the inspection to run UAST or
separate out java and kotlin checks.

Tests: existing
Change-Id: I54beb67bd3b1975c2a9bc26e0d7959f62d974f40

GitOrigin-RevId: fdb35d027cde6ffc10c1d468cac3499f55bda8fb