Switch insight prompt generation from issue to event
Prompt generation currently uses event from issue which is a sample
event. When the user navigates using the left/right arrow buttons, the
sample event does not change causing the prompt to be generated from an
event that is not the same as the one visible to the user.
Use app insight state's selectedEvent to generate the prompt which is
the same as the one visible to the user.

Bug: 338460277
Test: Updated
Change-Id: I063d6e416c6f6e203594ce89b9f8dfe66a80e471

GitOrigin-RevId: 18f89c99386a4405f4b6a8d1af900e7b58d7d132