Don't offer "disconnect" action when embedded LI is blocked on debugger
When Layout Inspector can't connect because the app is blocked on a
breakpoint, we offer two options "Resume debugger" to continue, and
"disconnect" to disconnect Layout Inspector.

The "disconnect" makes less sense in the context of embedded Layout
Inspector, as it should "un-toggle" Layout Inspector instead of just
disconnecting from the client.

For this reason this change removes the "disconnect" option for
embedded Layout Inspector.

Test: InspectorClientLaunchMonitorTest
Bug: 338931491
Change-Id: I5cda5e268de8a8938613a6ef8cf6406381af1501

GitOrigin-RevId: e4aab9ba0d667579552755ac20b7287fe4caceb6