Implement reset button for physical devices
A new Reset command and all setter commands are now requests with
a separate response. Each response has a new field: originalValues
which is true if all the UI settings values are the same as when
the first request to get UI settings was made.

The new Reset button will reset the values to what the values were
originally.

Bug: 308982261
Test: Update unit test for DeviceUiController
Change-Id: Ie1e9ee27525f478834319d289189b49e831ac5db

GitOrigin-RevId: ebf33c730551cc24e6b25a9a38d9e8e67eb50fe3