Simplify Compose autocomplete handling
Per discussion with folks on the Compose team, we're removing special
logic that adds named parameters for required parameters on Compose
functions. Named parameters should only be used for optional paramaters;
but since we don't add those in, there's no longer any need for a lot of
the specialized logic in place.

What remains is some logic around required lambdas as the last
parameter. Since these are very prevalent in Compose and are part of the
core functionality therein, we will still add an empty lambda for any
such parameters (which differs from Kotlin defaults).

Other than that, we now are relying on the standard Kotlin function
insertion. This should help reduce further bugs in the area and have
more consistency with Kotlin overall.

This change is also updating the text of the related setting for this
completion handler. There have been multiple bugs/feedbacks that
indicate the current text is unclear (in that users have no idea what it
does). Now that the setting only controls a single behavior and is quite
clear, updating it will help reduce confusion.

Fixes: 338071849
Test: ComposeCompletionContributorTest
Change-Id: I2d14a3aac3e1a06133add19f247e2eae45e877cd

GitOrigin-RevId: 58d076e2d25c7d4264ff3af635b9812d8dc4a9d9