Add check for startup trace event before task tab launch
...to prevent the task tab from being launched and entered when
the AndroidProfilerLaunchTaskContributor fails to execute the
underlying start trace command. If the user attempts to perform
a startup task, and the underlying start trace event is not found,
then an error notification is shown.

Included in this change is a refactor of the existing notification
code so that it can be shared with the "task" notifications. Also,
for consistency, the error dialog used for the launch task failure
was replaced with a new "task" notification.

Note: because the error dialog api is no longer used, it was deleted.

Fixes: 339256730
Test: added + updated + manual
Change-Id: I3fc649a272acbd0e0747fb7d249701432e39bcd5

GitOrigin-RevId: f5d787c7f83135a2ddd02fade37bc4044d6d2973