Update polling for WHS
The panel now has no visibility. It can be either created and running or
just disposed.
Because of that, we can remove the concept of "hidden". This will have
the side effect of an update of the state running immediately after
creation without any delay, producing a more responsive UI.

This change also updates the polling to be configurable via StudioFlags
and reduces it to every second.

Test: //tools/adt/idea/wear-whs:intellij.android.wear-whs_tests
Bug: N/A
Change-Id: If1402013ddb7eb5092790e27989f4bb2c2416dc9

GitOrigin-RevId: cdccbcd8f539bd469270014a69c0bbbf1ce168ea