Inline "Add * library" links for missing classes errors
When reporting missing classes errors in the issues panel, we add a link
to include the missing dependency in the project. However, the choice of
HTML was causing some confusion, because we add a line break after "("
and before the text ("Add * library [...]") with the link to add the
missing dependency.

This change fixes the formatting by removing the extra line.

Bug: 339615817
Test: Visually verified HTML render. RenderErrorContributorImplTest
 should still pass
Change-Id: I611a1ab947785a08ba154c18c386b8e80cbbc27d

GitOrigin-RevId: 33d2dc15fdc73c4026aecff192c3244b8862a6e5