Add equivalence check of source PSI for context from Preview parameter
Call argument resolution returns light class element, which is either
ULC (K1) or SLC (K2). Alas, as stated at KT-68158, SLC does not consider
source PSI equivalence in its `isEquivalentTo` implementation.
This adds a workaround by doing that directly at the check.
(This workaround can be removed once KT-68158 is fixed as we preferred,
or it can remain as-is, as the source PSI equivalence check itself is
still valid.)

Bug: 339683843
Test: WearTilePreviewContextComesFromParameterTest w/ and w/o K2
Change-Id: Iafe3400e54939297dcefe22fdc5192dbe244edab

GitOrigin-RevId: 56c1c53ab699efb2b0705055086f53ac60fc4dfa