Only show reasonable storage units
Nothing that this is used for actually supports byte-level granularity.
mksdcard will silently round down to the nearest KB boundary.

Terabytes of RAM / heap are not realistic any time soon.  The maximum SD
card size is 1023 GiB, and while internal storage could plausibly be >1
TiB, specifying it in GiB is reasonable for that rare case when someone
actually wants an emulator to store terabytes of data.

Bug: n/a
Test: n/a
Change-Id: I138852474ecdc88bb11fb10533514dee8c1ed077

GitOrigin-RevId: 65a5e0ba7069d04382f364747a1d3a125259687a