Make sure crash data has no duplicate keys
Currenty, if the crash data has duplicate keys, Android Studio will fail
when preparing the crash report
(AndroidStudioErrorReportSubmitter#getPlatformErrorData tries to build a
a map from a list of key-value pairs, and the map builder doesn't accept
duplicate keys).

To fix this issue, we will now append IDs to the duplicate keys' names.

Bug: 268521501
Test: Manual
Change-Id: I6cdfd33fef54654d91ba4096db80b4b335058aec

GitOrigin-RevId: ed2b4580654be9931814290b4051e8f92b99e2b8