Fix flaky test in WHS
The flakiness was caused by the polling that calls
WearHealthServicesStateManagerImpl#updateState which in turn updates
ongoingExercise based on the active exercise in the device manager. In
my previous implementation, I only overrode the exercise within the
state manager, meaning it could be set back to "false" before we would
call the reset method, leading to flakiness.

In this change, instead we override the value directly in the device
manager.

Fixes: 340913130
Test: bazel test --config=remote --runs_per_test=200 \
  //tools/adt/idea/wear-whs:intellij.android.wear-whs_tests
Change-Id: I436f2e0d2c8bc5edcad4dd9e8bb6fc4cc6f32c63

GitOrigin-RevId: 881551c8a9a834308c820e195701e47ea6f6025c