Cut down on the messages used for UI settings in agent
Instead of having a message for each type of change, create a request
for all the possible UI settings changes.
This cuts down on the number of messages for UI settings from 12 to 5.
It also makes the work of adding one more setting a little easier.

Also: Change the order of the settings such that the order is
consistent among the messages.

Bug: N/A
Test: N/A
Change-Id: I29a52a2d31710358eaf5e0bf047376d57304797c

GitOrigin-RevId: 5453f9096aac4c0f82c0f893cb7679ae684c03fa