Introduce AnimationTab to simplify SupportedAnimationManager
This change introduces a new class, AnimationTab, which encapsulates the
UI elements for an animation opened in individual tab. This simplifies
the SupportedAnimationManager interface by replacing the tabComponent
and addTimeline methods with a single tab property of type AnimationTab.
This change also updates the ComposeSupportedAnimationManager and
AnimationPreview classes to use the new AnimationTab class.

Bug: n/a
Test: Open inspector, add/remove tabs, verify functionality.
Change-Id: Iaff8668df53404b1083970a975380cfd4b9bbe19

GitOrigin-RevId: 6d7c685416a8b9146edbd8fb2033379e9452ef13