Add strongSkipping flag to Compose compiler plugin
When we specify `kotlin = "2.0.0-RC2"` (or 2.0.0-RC3) for Compose
compiler plugin version in the app `libs.versions.toml` file, it passes
`strongSkipping=true` compiler option by default (out of user's
control). The option is only defined by 2.0.0-RC, but it is not
available in the compose compiler plugin used by Studio. However, when
the Studio loads the compiler plugin, it uses the same compiler option
as the app. Therefore, if the app compiler option has
`strongSkipping=true`, Studio applies the same option when loading the
compiler plugin. Since the Compose compiler plugin used by Studio is
older version, and it does not have the option, it causes an exception.
To avoid the exception, this commit adds a patch to add the option to
the plugin.

Bug: 341233001
Test: Manually tested Jestcaster app with 2.0.0-RC2 Compose compiler
Change-Id: I4f4bf7e5b514b754574963f8e75deebb085b3608

GitOrigin-RevId: 3c2b4f6ae0de5628429c726697a072a562b1af8e