Remove extraneous editor team OWNERS
These OWNERS files were added for two reasons:

1. To help route test failures to appropriate components.
2. For directories where the editor team commits a lot of code, ensure we have ability to commit without extra approvals.

The second was needed because there was a plan to restrict directories without OWNERS to a subset of leads. That's no longer the plan, so those are no longer needed. This change leaves the test routing logic, but removes editor OWNERS where it's not really needed and is proving to be overly restrictive to other parts of the team.

Bug: n/a
Test: n/a
Change-Id: I77ac53b3ba2c7ad172979981da2ba662764d1470

GitOrigin-RevId: 8d2ce97067c658062d8f375dec89bd0f35034d51