Adjust compose completion for single param
When there's a single optional lambda parameter for a function, the
Kotlin plugin inserts two different autocomplete entries (one with, and
one without the labmda).

Our Compose autocomplete logic that specially renders Composable
function parameters only modifies one of these, and we end up with two
identical-looking completion entries that give different results.

The whole purpose of the custom logic for Composables is to handle cases
where functions have too many parameters to be easily viewable in
autocomplete. But that means it's not needed for functions with one (or
fewer) parameters. Simplifying this by just deferring to the Kotlin
plugin in these cases.

Note that this work made b/342188544 apparent, where the custom
rendering is effectively stripping out the package name for a function
that is otherwise added by the Kotlin plugin. This change won't affect
that functionality for functions with 2+ parameters, but the other bug
will track that.

Fixes: 339643347
Test: Covered by the existing modified tests.
Change-Id: I63a8b259cd317b3057c2de54c827a8267a83437c

GitOrigin-RevId: a98a99ef15b44e06f7a41bbb7df08ee4192d50dd