Use "_build_only" sdk's for studio UI tests
Build failures in the qa-fast target are observed due to an exessive
number of input files (70,609), exeeding the max limit of 70,000. To
address this, UI tests utilizing SDKs are being switched to _build_only
SDKs. This change will reduce the number of input files, as _build_only
SDKs exclude the generation of "/data/res**" files.

Bug: 340958820
Test: UI Tests
Change-Id: I8f7996db219d6e27d9595433952c6a1ff30678c7

GitOrigin-RevId: 735f426cfeec1b02e2a079dbca56214924c42c22