Change task tab title to be task name + tool name
The task names were converted to be in a title + subtitle
format in a past change, where the title was the task name,
and the subtitle is the underlying tool name (if there is one).
For example, system trace was renamed to be "Capture System
Activities" with a subtitle of "System Trace" for the task
grid items. Yet, the task tab itself never adapted to this
change, still showing only the underyling tool (e.g. System
Trace).

This change adapts the task tab to use the new task names,
with the tool name in parentheses if there is one. For example,
for Callstack Sample it will be "Find CPU Hotspots (Callstack Sample)"
and for Java/Kotlin Method Recording it will be
"Find CPU Hotspots (Java/Kotlin Method Recording)", but for a task
that does not have an underlying tool/subtitle, like Live View,
it will be "View Live Telemetry" (just the task name).

This change also features some small refactors to re-organize the
task based constants to separate out the utility methods, making
the connection to its corresponding test file clearer.

Fixes: 342233235
Test: added + updated
Change-Id: I042e7263cac94f1b549d94d5e9b4bd1efef143d1

GitOrigin-RevId: cc16139ada2ca715ce401fefeb89006c74565b36