Add call to stop ongoing task before launching a new one
This change makes it so that when launching a new task (via home
or past recordings tab), if there is a task tab already opened,
the user will be warned that launching a new task will...

  1. If the current task is ongoing -> terminate the current task
     and then start the new task, which will take over the task tab.
  2. If the current task is terminated -> start a new task, which
     will take over the task tab.

The aforementioned "warning" is facilitated via a "OK-Cancel" dialog
pop-up, with a "Do not ask again" checkbox for user convenience.

By implementing this logic, we fix the launch task error that
occurs when importing a task while a task is ongoing.

Fixes: 342105135
Fixes: 328099953
Test: added
Change-Id: I5ef05c0312f1704a6a013bcf7ccceec48d09995d

GitOrigin-RevId: 4e336564ff04c29ebaae21259652884ce33508b1