Introduce BuildTargetReference
An entity which represents a reference to a specific build target as
defined by the underlying build system and which is supposed to replace
usages of `AndroidFacet` in the context of rendering.

Start propagating its usage from `AndroidFacetRenderModelModule` where
it is intended to be consumed.

For now, stop propagating at the depths of a few callers and instantiate
a reference via `BuildTargetReference.gradleOnly()` factory method,
which represents the current behavior in all project systems.

Continue propagating the reference to more callers in upcoming changes
until in reaches the point in compose and layout previews, where a
reference to a specific build target can be constructed (i.e. both the
facet and a virtual file representing the build target are available).

Leave other usage intact (i.e. `gradleOnly()`). This won't change the
current behavior meaning that in non-Gradle project systems the result
of rendering may be incorrect.

Bug: 331381702
Test: n/a
Change-Id: Idd3fc59f77bee5d4876b8c0451e4abb9bc6e1404

GitOrigin-RevId: 8511633923f832f48bb6ff6eb1118bcd630a790e