Get project info for troubleshooter through a project system token
The relevant information is somewhat project system dependent, and
although the module-based predicates look nice and generic, they are
in fact not desperately so.  Move the current logic to the Gradle
Project System, accessed through a token (even though the designer
module currently uses the Gradle Project System module directly), and
provide a default implementation to (mostly) preserve behaviour on
other project systems.

The informational string generated is slightly different in form on
non-Gradle project systems as some fields are specific to the current
Gradle project / module model.

Making this project-system specific exposes the minimalism of the
AndroidProjectRule.inMemory test project rule.  Add a test and adjust
expected outputs.

Bug: 147586218
Test: updated and added
Change-Id: Ibf76b9cc026a44e83f42e45de36adb35f8ae8316

GitOrigin-RevId: 829b1fe20ca93288863525d757f193596cb6b827