[Sync] Remove RES_FOLDERS_RELATIVE_PATH property
It, along with TEST_RES_FOLDERS_RELATIVE_PATH, were used only in one
place.  That place can now read from the appropriate SourceProviders
directly, so these properties are now redundant and can be
removed.  (This should not affect deserializing older facets as I
believe that additional properties in the serialization are dropped.)

Fixes: 141909881
Bug: 246530964
Test: existing, snapshots updated
Change-Id: I68928b6e817c4a2bc95a33e1951cef97d9abe980

GitOrigin-RevId: 0dc31fd45e9c390165f412b0e37fb984eaa9cd39