Do not leak disposables in VisualLintService
When adding the issueProvider to the visual lint issue model, we
register a disposable to remove it if the surface is disposed. However,
when we manually remove the issueProvider, we leaked the disposable as
it is attached to the surface, and would only be handled when the
surface is disposed.
This ensures that, when we remove the issue provider, we also dispose
the disposable to reclaim the memory.

Fixes: 342649671
Test: LayoutEditorMemoryUseTest.addControl
Change-Id: I90ace854f93dd2340647bc57216b91e1ec68b5dc

GitOrigin-RevId: 9e8a73839d0ddbfd283d5258b79334b2a3ab7b84