Revert "Log error when safe args accessed in dumb mode"
This reverts commit d8d7c02ccecc883f0d1f079a34a991eca99733bf.

It turns out the logic in question is accessed frequently during dumb
mode, so removing the check is infeasible without a large rewrite. Given
that the fetcher exposes a ModificationTracker that is incremented
anytime dumb mode is started or stopped, any usages should be fine as
long as they use that tracker.

Fixes: 338287184
Test: Included
Change-Id: I66653dd50625cab45b7dd88f61c0bcffcdd83e78

GitOrigin-RevId: c05aed27fd75bd30d379e1538688ec48055164b5