Handle complex changes in LE invalidation
This changes the Live Edit group invalidation flow to utilize bytecode
analysis to determine the set of groups that need to be invalidated.
Additionally, this allows Live Edit to detect when changes have been
made that cannot be handled via group invalidation - namely, non-Kotlin
changes, changes to Compose restart lambdas, and changes to Composable
lambdas that are instantiated in non-Compose contexts.

For a full discussion of the design and implementation of this CL, see
go/live-edit-bytecode-analysis

Fixes: 336976336
Fixes: 338257199
Fixes: 318342040
Fixes: 311052100
Fixes: 337172924
Fixes: 216778745
Fixes: 216778745
Test: Existing, Manual
Change-Id: I7d56bde65fb68a6653fb9111ee34a67a5097e2fc

GitOrigin-RevId: 9a60aaac37f1c981d5fdb5a3b37a7408a408e3ae