Persist code folding state through IDE restart
This logic already exists in IntelliJ's default text editor provider,
but we were not appropriately calling it.

The change here calls read/writeState via the derived
PsiAwareTextEditorProvider class instead of the super
TextEditorProvider. Doing so runs all the same logic as before, with the
addition of saving code folding state.

Fixes: 201790861
Test: Manually verified
Change-Id: I02414009cc0f020620c796e0cc8e14a414dfd5cb

GitOrigin-RevId: 5a74edb39003691f2fc15833f76fc0d8172593f0