Add task in progress indication to tool window icon
This change adds the feature already present in Session-Based UX
where if there is an ongoing session, the tool window icon will
have a green icon indicating so. Also, if the user minimizes the
tool window, a little balloon will appear letting the user know
the session is in progress and that to end it, they should open
the tool window and press the stop button. For this change, we
do the same, with adapted terminology for the Task-Based UX.

Fixes: 343639873
Test: manual
Change-Id: I36f3b8f58afe407c058d7f8182b90338ed396f7b

GitOrigin-RevId: 740d7db6a29cbd63130ca32f944ab8fb94c1876a