Make non null cpu config check for new recordings only
A recent change (I58336601d9bef0e9935f76b2c9532f93947d7627)
added an assertion for a non-null cpu config before
entering a stage. The System Trace task was the only one that
could potentially return null, and it would do so in the case
that there was no device selected. Well, if the user imports
a trace file before opening the profiler tool window (assuming
no physical device plugged in), there will be no device selected,
and thus getCpuRecordingConfig did return null. However, this
check for a non-null config is only necessary for when the user
is capturing a new task recording, not when importing or opening
a past recording. Thus, the logic was adapted to only do the
check for a non-null config for when the user is capturing a
new task recording.

Fixes: 343989532
Test: added + updated + manually verified import case
Change-Id: Iec240ae95e7c958beeb2712a4aab1db9634a6a9e

GitOrigin-RevId: e228f07b64d4c27bae2a519b621de2176eb60ea5