Disable proceeding from profileable warning dialog
Before this change, if you invoked a run as profileable action
on api < 29 you would get prompted to use debuggable instead.
If you chose to continue, the resulting process had no guarantee
of being java-debuggable, leading to java-dependent tasks (e.g.
Java/Kotlin Allocations) failing. Because of this, it was decided
to not allow the user continue via this dialog, but rather redirect
them to use the appropriate, supported action for their device's
api level.

By doing the aforementioned change (not allowing user to proceed
as debuggable), we effectively would be not allowing the user to
perform startup tasks with a userdebug device or with a project
with an agp version below the minimum. This is because, on start
of a startup task, the profileable action will be invoked, and
because the criteria will fail to be met, the prompt from this
change will prevent the user from proceeding with the build, thus
preventing the startup task. To resolve this issue, the criteria
in the ProfileBuildAndLaunch api now also checks for userdebug devices
and supported agp projects when determining the build action, and
will use the debuggable action if the criteria is not met.

Fixes: 339258868
Bug: 279632659
Test: updated
Change-Id: I2e06b2db7d7afa10179873af4fe3daeeca06ea2c

GitOrigin-RevId: dce7b0a469e2253226cafb4079732779cada94e1