[Ps] Special handling of kotlinx coroutine -native-mt versions
The versions of some older kotlinx coroutine artifacts include,
somewhat like guava, a marker indicating a distinct artifact type.
Versions including the string "-native-mt" should not be treated as
preview versions, but rather as pseudo-stable versions in their own
series.

This is mostly historical (there haven't been -native-mt releases
for a couple of years, but older projects might well run into this, so
adjust the Project Structure Dialog's recommendation calculation to
include or exclude, as appropriate to the currently-used version,
coroutine versions including -native-mt versions.  (If "-native-mt" is
present in the version string, the underlying version will be treated
as a preview version, so we need to return early with a later
"-native-mt" suggestion if one is appropriate.  If it is not present,
then we need to fall back to the usual version suggestion algorithm,
excluding any "-native-mt" versions from consideration).

Bug: 211949862
Test: existing, manual
Change-Id: Icc400cb6ea26d2400af456f1236b452ceccbfdb9

GitOrigin-RevId: 6aeca3f8203310776762e8cbce8e2ee73e90dae7