Move GradleFiles service to Android Studio plugin
Recently Intellij added a fix to address IDEA-328075 issue about slow
operation, however, this caused sync banner to not be triggered
when modifying the build files after opening the file while sync
didn't finished. This is because, the following check:
CommonAndroidUtil.getInstance().isAndroidProject(myProject)
requires the project to be synced, meaning that no PSI events will
be received for all the opened file while the sync is in progress,
requiring users to reopen the file subscribing again the
GradleFileChangeListener on PsiManagerImpl.myTreeChangeListeners.

Since GradleFiles service is strictly linked to studio, this one was
moved to only be added for android studio.

Fixes: 342985434
Fixes: 335735306
Test: Existing
Change-Id: I9849271efe47bc841cfda8822232ae36fc650d26

GitOrigin-RevId: 1babe3aaf0732d15485719728e76fdf7111b4b42