Maybe improve addLibraryDependency a bit
We have to join two sources of data (from sync and from the IDE's
OrderEnumerator) by something that is approximately a String key,
where the format of the string is slightly different on either side.

Joy.  See also:
https://web.archive.org/web/20030709164712/http://www.advogato.org/person/wnewman/diary.html?start=6

We can do this for more than (plain) Java libraries, though: our
new (2021-era) taxonomy allows us to check IdeArtifactLibrary for the
artifactAddress.  This address is nearly the same as the
OrderEnumerator's names, except for the external system prefix, which
we have already encountered in the Gradle Project System as something
which needs to be accounted for.

Once this is addressed (no pun intended), the platform action for
adding the needed dependency to the module is more functional.  I have
no idea why it is labelled `add library 'classes' to classpath` but
that is left as an exercise for the truly devoted reader.

Bug: 339008412
Test: manual
Change-Id: Ia22d2b7c49fa6291d0885566ac6014fec7c2f0f3

GitOrigin-RevId: 24de04768b4de537d101f02128a654614ebf612e