Treat null PSI state as a new file
When a new file is created in Studio, the document change callbacks run
before the editor is opened, meaning we'll process a change with no base
PSI snapshot. When we see a null PSI state, assume it is because the
file was newly created, and log.

If we get into a weird state where there is no PSI baseline for some
other, bug-related reason, this will only erroneously permit *some*
disallowed changes on the *first* Live Edit; we will subsequently still
catch the full set of incompatible changes. Thus, it's better to be
permissive here.

Also fixes a bug with PSI validation where we never reset the PSI
basline and thus PSI validation errors will persist even after the app
has been re-run.

Fixes: 339587836
Test: Manual
Change-Id: Iccba09321106fd3298486b1d99a224382443beb7

GitOrigin-RevId: 23d7e792e2a6847fd36223aa069a9e145320ff75