Make sure selections are valid before fetching error
When fetching errors for specific task starting points (e.g. startup
task errors), it assumes all selections made are valid. Thus, an
assertion was added to enforce such assumption. Also, the function
for getting starting point specific errors is now only called within
the greater error fetching function so that it can first check for
valid selections.

To fetch the startup specific errors now, the general getStartTaskError
method is called with the PROCESS_START starting point hard-coded. This
approach is much safer, as it can leverage the aforementioned checks
for valid selections already made by the method. If there is a selection
error, it is filtered out and not shown; if all selections are valid,
and the error is startup specific, it will show still.

It is worth noting that this change does not functionally or visually
change anything, just makes the order of selection validation safer.

Bug: 277797248
Test: existing
Change-Id: Ib190c05f9f9bf6c8a776b0e9ed95e5c7f291c615

GitOrigin-RevId: 660645dc7886b516daf352dbd31ec431564ad6ce