Make projectPsiTreeObserver non-nullable
This "observer" is different from other "observers" in this file, in
that it doesn't register itself for any platform events. Rather, it's
called external whenever the psiListener property returns it.

Since the deletion/recreation of this object doesn't do anything other
than control when it's returned, we can avoid the churn and needing to
make changes within a lock by just checking whether we have any
listeners at the time the property is accessed.

Bug: 343653602
Test: ResourceNotificationManagerTest
Change-Id: I0256c5bf2b6e7161f602584795199a089e93dfc8

GitOrigin-RevId: 589fa1280c320c01914dbf2a93a6b6845284f85d