Wait for the palette to properly initialize.
In this UI testcase a click on the wanted category is made, but
the click was made before the palette was properly initialized.

The click event was created while the category list was very wide,
but the click event was fired after the palette doLayout() was
called where the category list width was reduced to about the half
width (giving space to the item list on the right).

The click then missed the category panel and we get a strange error
about "Common" category was selected instead of "Buttons".

Fixes: 343985202
Test: ComponentSelectionSplitViewTest
Change-Id: If9b10deb9c8efaad42b0b9eb8413076e09235ea2

GitOrigin-RevId: 2bbb16dd72adf58cf8fea4076b20aafede104e39