Use withAndroidModels() in preview-related wear tests
The use of withAndroidModel() with its heritage from default android
projects leads to quite an unrealistic project setup: which is then
compounded in these tests by needing to add additional source roots by
hand.

Instead, use withAndroidModels(...) with an android application in the
project's root module.  This is still a little unrealistic, but at
least comes with all the relevant source sets.  However, although the
module content roots are where they are expected, the sourceSets are
in java/ subdirectories, so move the test files to where they are
actually in those sourceSets.

This affects not just source files in the tests themselves, but also
the wear tile preview annotation stub file.  Placing it in a java/
subdirectory should not change behaviour for the default android
module model, while allowing it to be picked up from more realistic
project setups.

Bug: none filed
Test: test changes only
Change-Id: I88770354380994c9a2cdb7799a72d0902f84baf6

GitOrigin-RevId: d6101d7205c6cb1f525fbadd392df58497685aff