Keep the user's edits during updates
The properties panel can be asked to update the values based on
external changes and resolvers that are loaded delayed.

If the user is editing during one of these updates, keep the user's
edits. Before this change the modifications were lost during updates.

CommonComboBox is using a CommonTextField for its edit field.
When a value is chosen from the combo box we do NOT want to keep the
value in the edit field. The solution chosen here is to make a
callback to make the combo box update the text value in the edit field
before updating the property value.

Fixes: 341537014
Test: Added unit tests
Change-Id: I6c0ffb9dfe003cf37d6cfcc53054cad303d53c0e

GitOrigin-RevId: d0e96a46e949b5074baeedcf8f94d32bfbfee02e