Handle split mode when resizing tool window
Tabs in the running devices tool window can be in split mode, where
multiple tabs are visible at the same time.

In this mode there are multiple content managers, each one handling
its own content. Therefore getting the width using
toolWindow.component.width does not work, because the first component
is empty.

This change uses the largest component in the tool window as the
point of referece for the current width.

Test: NA
Bug: 347072130
Change-Id: Ic3406b85834557dc6698c82dfc3a29e714c60eda

GitOrigin-RevId: 93f0e397b7f9c6f875cf0b8b2af52e113689467b