Add code links when there is an invalid context usage
This CL adds code links to the problems panel so the user can navigate
to code that might be causing the issue. The links are taken elements
from the stack trace that might be of interest to the user. The code
links reference stack trace elements from code that is part of the
tile preview's source file.

Fixes: 346819620
Test: WearTileRenderErrorContributorTest + manual
Change-Id: I3060596e07157c1d429b1d520dad00f58864c33b

GitOrigin-RevId: 6a1cf1b2bff7128986c54abe60c4abdce2e54417