Add baseline profile plugin details to AUA for AGP upgrade
When AGP is being upgraded to 8.3.1 or above, and the baseline profile
plugin is also present at version 1.2.2 or below, sync fails because the
versions are incompatible. This adds functionality so that when AGP is
being upgraded in this situation, the BP version also gets upgraded to
1.2.4, where there is no longer a sync error but only a warning for any
further incompatible versions.

Bug: 346548760
Test: manual, added
Change-Id: I3f9e0dfb6871b73b42316dc5f3919544728f089d

GitOrigin-RevId: 4baf7798041cd4605cbb1a3c97e73edd47251b91