Fix flaky ComposePreviewRepresentationTest
This change fixes flakiness in ComposePreviewRepresentationTest.

There were three causes of flakiness:
1. The initial cause was a NlModel that was attempted to be registered
to a parent disposable which was already disposed. This was happening
due to the following scenario:
  a. refresh request starts to be processed
  b. we check if ComposePreviewRepresentation has been disposed, it
     hasn't been yet, so we proceed
  c. the test finishes and the root disposable is disposed
  d. the refresh requests gets to
     NlDesignSurface.updatePreviewsAndRefresh and creates the NlModel.
  e. The NlModel.init method registers itself to the now disposed
     parent

This change fixes this by making the NlModel.init call tryRegister
instead. Another solution, that would have only solved
ComposePreviewRepresentationTest, would have been to first check that
all pending refreshes are finished before letting the test finish.
However, this change went with the current approach as it should be more
resilient to other scenarios.

2. a concurrent modification exception occurring only in the test
testUiCheckModeWithColorBlindCheckEnabled. This was happening 1-2 times
per 1000 runs. This was originating from DefaultTreeLayoutCache.kt:1075
when calling ProblemsViewToolWindowUtils.addTab. This seems to be due to
the call not being made on the EDT. Moving the call to the EDT seems
to resolve the issue.

3. A concurrent modification exception occurring in
TestProjectSystemBuildManager. This was happening 1-2 times per 1000
runs. The issue was because ComposePreviewRepresentation.activate adds
build listeners. TestProjectSystemBuildManager.buildCompleted would
sometimes still be running when those listeners were added. In that
method we iterate over the listeners which led to the concurrent
modification.

This change went with the approach of creating a copy of the listeners
before iterating over them to avoid any concurrent modification.
Another approach considered was using a lock to guard any read/write
on that list. This CL goes with the former approach instead of the
latter to avoid any potential dead lock if a listener tries to add
another listener while already in a lock.

Fixes: 344852674
Test: bazel test --config=remote --runs_per_test=1000\
  //tools/adt/idea/compose-designer:\
  intellij.android.compose-designer\
  .tests_tests__ComposePreviewRepresentationTest
Change-Id: I927a5ee4f9b90142adf12708d505dbd2b560be74

GitOrigin-RevId: 8f45a5cff76805a27fc1ddd2ec1c41bdf787d965