Remove old StackTraceAnalyzer code
We were using it solely for the purpose of collecting metrics about
the confidence of line matches. I don't think we are using these
metrics in any meaningful way so they can probably be safely removed.
In addition now we have CVS based line matching, we should collect
metrics there if necessary.

Test: n/a - refactoring change.
Bug: n/a
Change-Id: Ie030260729aef7cc4dae89450139c1fa64dae1e6

GitOrigin-RevId: d1659a8a0ba9cb247184418418bdfbb4ee5036c3