Add separate settings for: Samsung, OnePlus, OPPO, Xiaomi
It turns out that some Samsung devices are sensitive to the global
setting: navigation_bar_gesture_while_hidden being:
- 1 for gesture navigation
- 0 for button navigation
Some Samsung devices also requires the correct overlay to be selected.

Xiaomi devices seem to be sensitive to the global setting:
force_fsg_nav_bar being:
- 1 for gesture navigation
- 0 for button navigation

OnePlus & OPPO devices are sensitive to the secure setting:
hide_navigationbar_enable being:
- 3 for gesture navigation
- 0 for button navigation

This change modifies both the way gesture navigation is detected and
the way it is being set.
For Xiaomi the default detection of selected overlay did not work,
since the overlays seem to be ignored in the Xiaomi implementation.

Fixes: 347800946
Fixes: 348404170
Test: N/A
Change-Id: Ib47296d8981861c9bc98bc4535ee61e13379bec8

GitOrigin-RevId: 98f09503a036dcdc85f27f222da69c7fd4265701