Rename the label for gesture navigation to Navigation Mode
Also rename the choices in the dropdown to:
- Gestures
- Buttons

Since on some devices we may get a 2 button mode and for others
if may be a 3 button mode. The difference between the 2 or 3 buttons
makes no difference for the user's Application - thus there is no
need to offer both button choices.

Bug: 347776326
Test: N/A
Change-Id: I835a4933faf23d447eb884b07d3dd5cf2b0d3bd5

GitOrigin-RevId: 327908e088b3d34520a19b70228a938e20e9d46d